fix: int32

+++ app/src/main/java/com/aicane/app/data/remote/dto/destination/CreateDestinationRequest.kt
@@ -9,5 +9,5 @@ data class CreateDestinationRequest(
     val targetText: String,
     val latitude: Double,
     val longitude: Double,
-    @SerialName("radiusM") val radius: Double,
+    @SerialName("radiusM") val radius: Int,
 )
